requirement.txt
PyQt5
biplist
protobuf==3.20.1
silkpy-1.0.0-cp38-cp38-win_amd64.whl
pyaudio
mmkv-1.0.0-cp38-cp38-win_amd64.whl
lxml